[FIX] UIHV: Writing data to file
Issue:
Previously, there was a dynamically allocated buffer of a
constant size that have been used for all of the UI elements
regardless of their count and used data size.
Sometimes this have led to buffer overflow.

Solution:
Now buffer is not used at all, the whole data is written directly
to a file.

Change-Id: I847321db086f1c8e667d2ba7d12572fb29a6897f
Signed-off-by: Alexander Aksenov <a.aksenov@samsung.com>